8/25/2024 12:44 Software state
main.py is currently dysfunctional, code idles while saving jpg in Camera.py. Fix must be worked on OR an alternative approach to be found. SeeBlue_v2.py works, an alternative to using OpenCV must be found OR image color codes must be converted to binary and analyzed as such.